requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
Babel==2.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
blinker==1.4
bob.blitz==2.0.13
bob.core==2.1.8
bob.extension==2.4.2
bob.io.base==2.1.0
bob.io.image==2.2.2
bob.io.video==2.0.15
bob.ip.base==2.0.15
bob.ip.color==2.0.9
bob.ip.draw==2.0.8
bob.ip.facedetect==2.1.3
bob.ip.facelandmarks==1.0.4
bob.learn.boosting==2.0.12
bob.math==2.1.1
bob.sp==2.0.9
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
cdecimal==2.3
cffi==1.10.0
chardet==3.0.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
configparser==3.5.0
contextlib2==0.5.5
coverage==4.4.1
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
cyvlfeat==0.4.6
dask==0.14.3
datashape==0.5.4
decorator==4.0.11
distributed==1.16.3
docopt==0.6.2
docutils==0.13.1
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
gevent==1.2.1
greenlet==0.4.12
grin==1.2.1
guess-language==0.2
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.999
idna==2.5
imageio==2.2.0
imagesize==0.7.1
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
MarkupSafe==0.23
matplotlib==2.0.2
menpo==0.8.1+0.g0977044a.dirty
menpofit==0.5.0
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0+0.ge79330a.dirty
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
partd==0.3.8
pathlib==1.0
pathlib2==2.2.1
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
ply==3.10
prompt-toolkit==1.0.14
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
PyAudio==0.2.7
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.0.7
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.14.2
rope==0.9.4
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
statsmodels==0.8.0
subprocess32==3.2.7
sympy==1.0
tables==3.3.0
tblib==1.3.2
Tempita==0.5.2
terminado==0.6
testpath==0.3
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
Werkzeug==0.12.2
Whoosh==2.7.4
widgetsnbextension==2.0.0
wrapt==1.10.10
WTForms==2.1
-e git+https://github.com/dmlc/xgboost@8b827425b2e109410e843f88238b344089cddf4c#egg=xgboost&subdirectory=python-package
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
